Improve EQC tests and fix bug in histogram code
The histogram code was wrong in the case where a value was the same as
the maximum allowed. Before the fix, it'd get put in in the 'Bins + 1'
bin, instead of the final bin actually used in the histogram.

The EQC tests have also been significantly improved so that error
messages from R are captured, shrinking works correctly and the property
can be tested from quickcheck directly without involving eunit.