Non-deterministic test needs a little ?SOMETIMES
prop_quantile/0 is by it's nature non-deterministic, and the
comments/notes make it clear. Rather than tolerate it breaking the build
periodically, this commit changes the property by using eqc's
`?SOMETIMES` macro. If a test fails 3 times in a row then we assume that
it is a deterministic failure, a bug. Otherwise, assume it is the noted
"loosy goosy" nature of the comparison and pass the property. NOTE:
still think we should find and fit the error bounds.